add routes for series, layouts for series and articles
